Split extraction internals into focused modules.
Move types, config, media loading, error mapping, retries, agent
construction, sessions, and batch execution out of the former
_extract.py monolith so each concern lives in one place and the
duplicated URL, retry, and batch-sync paths share helpers. Public
extract APIs and behavior are unchanged.
