+++ src/openextract/__init__.py
@@ -2,25 +2,28 @@
 openextract - Extract structured data from documents, images, audio, and video using LLMs.
 """
 
-from ._extract import (
-    AsyncExtractor,
-    ExtractionInput,
-    ExtractionResult,
-    Extractor,
-    RetryPolicy,
-    Usage,
-    extract,
-    extract_async,
+from ._batch import (
     extract_many,
     extract_many_async,
     extract_many_with_results,
     extract_many_with_results_async,
+    iter_extract_many_async,
+)
+from ._extract import (
+    extract,
+    extract_async,
     extract_with_usage,
     extract_with_usage_async,
-    iter_extract_many_async,
+)
+from ._session import AsyncExtractor, Extractor
+from ._styles import ExtractionStyle
+from ._types import (
+    ExtractionInput,
+    ExtractionResult,
+    RetryPolicy,
+    Usage,
     total_usage,
 )
-from ._styles import ExtractionStyle
 from .exceptions import (
     ExtractionError,
     InputTooLargeError,
