+++ src/openextract/_extract.py
@@ -1,1105 +1,80 @@
-"""Core extraction functionality."""
+"""One-shot extraction APIs and internal re-exports."""
 
 from __future__ import annotations
 
-import asyncio
-import ipaddress
-import math
-import mimetypes
-import os
-import random
-import shutil
-import socket
-import tempfile
-import threading
-import time
-from collections.abc import (
-    AsyncIterator,
-    Awaitable,
-    Callable,
-    Iterable,
-    Iterator,
-    Mapping,
-    Sequence,
-)
+from collections.abc import AsyncIterator, Iterator
 from contextlib import asynccontextmanager, contextmanager
-from dataclasses import dataclass
-from email.utils import parsedate_to_datetime
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, BinaryIO, Literal, TypeVar, cast, overload
-from urllib.parse import urlparse
+from typing import TYPE_CHECKING, TypeVar, cast
 
 import httpx
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
+
+from ._agent import (
+    Agent,
+    _build_agent,
+    _install_hint,
+    _model_identifier,
+    _resolve_run_inputs,
+    _route_model,
+    _run_extraction,
+    _run_extraction_async,
+    _usage_from_result,
+)
+from ._batch import (
+    _run_with_shared_agent,
+    _run_with_shared_agent_result,
+    extract_many,
+    extract_many_async,
+    extract_many_with_results,
+    extract_many_with_results_async,
+    iter_extract_many_async,
+)
+from ._config import (
+    _DEFAULT_RETRY_MAX_BACKOFF,
+    _max_redirects,
+    _resolve_max_input_bytes,
+    _url_fetch_timeout,
+    _validate_retry_options,
+)
+from ._errors import (
+    _extraction_errors,
+    _is_transient_model_exception,
+    _map_exception,
+    _model_retry_after,
+    _model_status_code,
+    _parse_retry_after,
+)
+from ._media import (
+    _fetch_url,
+    _fetch_url_async,
+    _get_media,
+    _get_media_async,
+    _get_media_type,
+    _is_public_ip,
+    _is_safe_host,
+    _item_source_label,
+    _read_from_path,
+    _read_url_with_client,
+    _safe_source_context,
+)
+from ._retry import _retry_delay, _run_with_retries_async, _run_with_retries_sync
+from ._session import AsyncExtractor, Extractor
+from ._styles import ExtractionStyle, normalize_style, prepared_style_run
+from ._types import (
+    ExtractionInput,
+    ExtractionInputLike,
+    ExtractionResult,
+    RetryPolicy,
+    Usage,
+    _resolve_item,
+    total_usage,
+)
 
 if TYPE_CHECKING:
     from pydantic_ai import Agent as PydanticAgent
     from pydantic_ai.models import Model
-    from pydantic_ai.models.instrumented import InstrumentationSettings
-    from pydantic_ai.settings import ModelSettings
-else:
-
-    def Agent(*args, **kwargs):
-        """Construct a Pydantic AI agent without loading it at package import time."""
-        from pydantic_ai import Agent as PydanticAgent
-
-        return PydanticAgent(*args, **kwargs)
-
-
-from ._styles import (
-    ExtractionStyle,
-    materialize_text_document,
-    normalize_style,
-    prepared_style_run,
-    style_capabilities,
-    style_run_inputs,
-)
-from .exceptions import (
-    ExtractionError,
-    InputTooLargeError,
-    ModelError,
-    ProviderNotInstalledError,
-    SchemaValidationError,
-    UrlFetchError,
-)
 
 T = TypeVar("T", bound=BaseModel)
-
-# A raw media source accepted directly by the public APIs: a local path or
-# http(s) URL string, an ``os.PathLike`` (e.g. ``pathlib.Path``), raw bytes, or
-# a binary file-like object with a ``.read()`` method.
-MediaSource = str | os.PathLike[str] | bytes | BinaryIO
-
-# A media source after ``os.fspath`` normalization: no ``os.PathLike`` remains.
-ResolvedSource = str | bytes | BinaryIO
-
-
-@dataclass(frozen=True)
-class ExtractionInput:
-    """A single input for extraction with optional per-item media metadata.
-
-    Wraps a raw :data:`MediaSource` so heterogeneous batch inputs can specify
-    their own ``media_type`` (and an optional safe ``name`` for diagnostics)
-    without falling back to a single batch-wide media type.
-
-    Attributes:
-        source: The media source — a local path, ``http(s)://`` URL,
-            ``os.PathLike``, raw ``bytes``, or a binary file-like object.
-        media_type: Optional MIME type for this item. Required when ``source``
-            is ``bytes`` or a file-like object and no batch-wide ``media_type``
-            override is supplied. Overrides inference for path/URL sources.
-        name: Optional safe source name recorded on :class:`ExtractionResult`
-            diagnostics. Never populated with raw content or credentials.
-    """
-
-    source: MediaSource
-    media_type: str | None = None
-    name: str | None = None
-
-
-# Anything accepted as a single input or batch item: a raw :data:`MediaSource`
-# or a structured :class:`ExtractionInput`.
-ExtractionInputLike = MediaSource | ExtractionInput
-
-
-@dataclass(frozen=True)
-class ExtractionResult[T]:
-    """Diagnostics-rich result of one extraction.
-
-    ``extract_many_with_results`` returns these so callers can account for
-    token usage, observe retries and timing, and record safe provenance without
-    retaining raw media, credentials, query strings, or provider internals.
-
-    Attributes:
-        output: The validated schema instance.
-        usage: Token usage from the successful model call.
-        attempts: Number of model-call attempts, including the initial call and
-            any :class:`ModelError` retries (always ``>= 1`` on success).
-        duration: Wall-clock seconds spent on this item, including retries.
-        model: The model identifier that produced the output, when known.
-        media_type: The media type requested for this item, when provided.
-        source: A sanitized source label (``ExtractionInput.name``, or a
-            credential/query-stripped path/URL context); ``None`` for unnamed
-            bytes/file-like inputs.
-        warnings: Extensible, currently empty diagnostics channel.
-    """
-
-    output: T
-    usage: Usage
-    attempts: int
-    duration: float
-    model: str | None
-    media_type: str | None
-    source: str | None
-    warnings: tuple[str, ...] = ()
-
-
-_DEFAULT_MEDIA_TYPE = "application/octet-stream"
-_URL_PREFIXES = ("http://", "https://")
-_DEFAULT_URL_FETCH_TIMEOUT = 30.0
-_DEFAULT_MAX_REDIRECTS = 10
-_DEFAULT_RETRY_MAX_BACKOFF = 60.0
-_DEFAULT_MAX_INPUT_BYTES = 50 * 1024 * 1024
-_INPUT_READ_CHUNK_SIZE = 64 * 1024
-_OPENAI_PREFIX = "openai:"
-_OPENAI_RESPONSES_PREFIX = "openai-responses:"
-_URL_TIMEOUT_ENV = "OPENEXTRACT_URL_TIMEOUT"
-_MAX_REDIRECTS_ENV = "OPENEXTRACT_MAX_REDIRECTS"
-_ALLOW_PRIVATE_URLS_ENV = "OPENEXTRACT_ALLOW_PRIVATE_URLS"
-_MAX_INPUT_BYTES_ENV = "OPENEXTRACT_MAX_INPUT_BYTES"
-_BYTES_MEDIA_TYPE_REQUIRED = (
-    "media_type is required when input_file is bytes or a file-like object; "
-    "pass it explicitly, e.g. extract(..., media_type='application/pdf')."
-)
-_TRANSIENT_HTTP_STATUSES = frozenset((408, 409, 425, 429))
-_PROVIDER_MODULE_NAMES: tuple[tuple[str, str], ...] = (
-    ("openai", "openai"),
-    ("anthropic", "anthropic"),
-    ("google", "google"),
-    ("botocore", "bedrock"),
-    ("cohere", "cohere"),
-    ("huggingface_hub", "huggingface"),
-    ("groq", "groq"),
-    ("mistralai", "mistral"),
-    ("grpc", "xai"),
-)
-_TRANSIENT_ERROR_NAMES = (
-    "timeout",
-    "connection",
-    "ratelimit",
-    "throttl",
-    "resourceexhausted",
-    "servererror",
-    "internalserver",
-    "serviceunavailable",
-)
-_PERMANENT_ERROR_NAMES = (
-    "authentication",
-    "permission",
-    "forbidden",
-    "badrequest",
-    "invalidrequest",
-    "validation",
-    "unprocessable",
-    "notfound",
-    "accessdenied",
-)
-
-# Exact (module, class) signatures for provider/model error bases we classify
-# as ``ModelError``. The classifier checks the exception's existing MRO instead
-# of importing provider SDKs. ``openai.APIError`` also covers OpenRouter and
-# Cerebras since both go through the openai SDK.
-_MODEL_ERROR_SIGNATURES = frozenset(
-    {
-        ("pydantic_ai.exceptions", "ModelAPIError"),
-        ("openai", "APIError"),
-        ("anthropic", "APIError"),
-        ("google.genai.errors", "APIError"),
-        ("botocore.exceptions", "ClientError"),
-        ("botocore.exceptions", "BotoCoreError"),
-        ("cohere.core.api_error", "ApiError"),
-        ("huggingface_hub.errors", "HfHubHTTPError"),
-        ("groq", "APIError"),
-        ("mistralai.client.errors.mistralerror", "MistralError"),
-        ("grpc", "RpcError"),  # xAI SDK uses gRPC; pydantic-ai may surface this directly
-    }
-)
-
-# pydantic-ai model-id prefix -> the openextract optional-dependency extra that
-# installs that provider's SDK. OpenAI-compatible providers (Cerebras, Ollama)
-# ship through the openai SDK, so they map to the ``openai`` extra. Prefixes not
-# listed here fall back to suggesting ``openextract[all]``.
-_PROVIDER_EXTRAS: dict[str, str] = {
-    "openai": "openai",
-    "openai-chat": "openai",
-    "openai-responses": "openai",
-    "anthropic": "anthropic",
-    "google-gla": "google",
-    "google-vertex": "google",
-    "bedrock": "bedrock",
-    "cohere": "cohere",
-    "groq": "groq",
-    "huggingface": "huggingface",
-    "mistral": "mistral",
-    "openrouter": "openrouter",
-    "xai": "xai",
-    "cerebras": "openai",
-    "ollama": "openai",
-}
-
-
-def _is_model_exception(exc: BaseException) -> bool:
-    """Return whether an exception inherits from a supported model error base.
-
-    Provider SDKs have already created the exception by the time this runs, so
-    their classes are present in its MRO. Comparing exact module/class
-    signatures preserves subclass handling without importing unrelated SDKs.
-    """
-    return any(
-        (error_type.__module__, error_type.__name__) in _MODEL_ERROR_SIGNATURES
-        for error_type in type(exc).__mro__
-    )
-
-
-@dataclass(frozen=True)
-class Usage:
-    """Token usage information for a single extraction call."""
-
-    input_tokens: int
-    output_tokens: int
-    total_tokens: int
-
-
-@dataclass(frozen=True)
-class RetryPolicy:
-    """Retry configuration shared by every call made through an extractor session."""
-
-    max_retries: int = 0
-    backoff: float = 1.0
-    max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF
-
-    def __post_init__(self) -> None:
-        _validate_retry_options(self.max_retries, self.backoff, self.max_backoff)
-
-
-def _get_media_type(file_path: str) -> str:
-    """Return the MIME type for a file path (e.g. 'application/pdf')."""
-    media_type, _ = mimetypes.guess_type(file_path)
-    return media_type or _DEFAULT_MEDIA_TYPE
-
-
-def _allow_private_urls() -> bool:
-    """Return True when SSRF host validation is disabled via env var."""
-    return os.environ.get(_ALLOW_PRIVATE_URLS_ENV, "").lower() in ("1", "true", "yes")
-
-
-def _env_positive_float(name: str, default: float) -> float:
-    """Parse a positive float from ``name``; return ``default`` when unset or invalid."""
-    raw = os.environ.get(name, "").strip()
-    if not raw:
-        return default
-    try:
-        value = float(raw)
-    except ValueError:
-        return default
-    return value if value > 0 else default
-
-
-def _env_positive_int(name: str, default: int) -> int:
-    """Parse a positive int from ``name``; return ``default`` when unset or invalid."""
-    raw = os.environ.get(name, "").strip()
-    if not raw:
-        return default
-    try:
-        value = int(raw)
-    except ValueError:
-        return default
-    return value if value > 0 else default
-
-
-def _url_fetch_timeout() -> float:
-    """HTTP timeout in seconds for URL fetches (``OPENEXTRACT_URL_TIMEOUT``)."""
-    return _env_positive_float(_URL_TIMEOUT_ENV, _DEFAULT_URL_FETCH_TIMEOUT)
-
-
-def _max_redirects() -> int:
-    """Maximum redirect hops when fetching URLs (``OPENEXTRACT_MAX_REDIRECTS``)."""
-    return _env_positive_int(_MAX_REDIRECTS_ENV, _DEFAULT_MAX_REDIRECTS)
-
-
-def _resolve_max_input_bytes(max_input_bytes: object) -> int:
-    """Resolve and validate the per-input byte limit.
-
-    An explicit value wins over ``OPENEXTRACT_MAX_INPUT_BYTES``. Invalid
-    configured values fail closed instead of silently disabling the limit.
-    """
-    value = max_input_bytes
-    from_environment = False
-    if value is None:
-        raw = os.environ.get(_MAX_INPUT_BYTES_ENV, "").strip()
-        if not raw:
-            return _DEFAULT_MAX_INPUT_BYTES
-        from_environment = True
-        try:
-            value = int(raw)
-        except ValueError as exc:
-            raise ValueError(f"{_MAX_INPUT_BYTES_ENV} must be a positive integer.") from exc
-    if isinstance(value, bool) or not isinstance(value, int) or value < 1:
-        if from_environment:
-            raise ValueError(f"{_MAX_INPUT_BYTES_ENV} must be a positive integer.")
-        raise ValueError("max_input_bytes must be a positive integer.")
-    return value
-
-
-def _safe_source_context(source: str) -> str:
-    """Return source context without URL credentials, query strings, or fragments."""
-    if source.startswith(_URL_PREFIXES):
-        parsed = urlparse(source)
-        host = parsed.hostname or "unknown-host"
-        try:
-            parsed_port = parsed.port
-        except ValueError:
-            parsed_port = None
-        port = f":{parsed_port}" if parsed_port is not None else ""
-        return f"URL {parsed.scheme}://{host}{port}{parsed.path or '/'}"
-    return f"path {Path(source).name!r}"
-
-
-def _input_too_large(*, limit: int, observed: int, source: str) -> InputTooLargeError:
-    return InputTooLargeError(
-        f"{source} exceeds the configured size limit ({limit} bytes); "
-        f"got at least {observed} bytes. Set {_MAX_INPUT_BYTES_ENV} or pass "
-        "max_input_bytes=... if this is intentional."
-    )
-
-
-def _enforce_max_input_bytes(data: bytes, *, limit: int, source: str) -> bytes:
-    if len(data) > limit:
-        raise _input_too_large(limit=limit, observed=len(data), source=source)
-    return data
-
-
-def _read_file_like_limited(stream: BinaryIO, *, limit: int, source: str) -> bytes:
-    """Read a binary stream in bounded chunks, including non-seekable streams."""
-    chunks: list[bytes] = []
-    total = 0
-    while True:
-        chunk = stream.read(min(_INPUT_READ_CHUNK_SIZE, limit - total + 1))
-        if not chunk:
-            break
-        total += len(chunk)
-        if total > limit:
-            raise _input_too_large(limit=limit, observed=total, source=source)
-        chunks.append(chunk)
-    return b"".join(chunks)
-
-
-def _is_public_ip(ip: ipaddress.IPv4Address | ipaddress.IPv6Address) -> bool:
-    """Return True only for globally routable public unicast addresses.
-
-    Treats IPv4-mapped IPv6 (``::ffff:a.b.c.d``) as its underlying IPv4 so
-    that e.g. ``::ffff:127.0.0.1`` is correctly classified as loopback.
-    """
-    if isinstance(ip, ipaddress.IPv6Address) and ip.ipv4_mapped is not None:
-        ip = ip.ipv4_mapped
-    return ip.is_global
-
-
-def _is_safe_host(host: str | None) -> bool:
-    """Return True when ``host`` resolves only to public IP addresses.
-
-    Refuses missing/empty hosts, IP literals in private/loopback/link-local/
-    multicast/reserved ranges (including AWS/GCP metadata at
-    ``169.254.169.254``), hostnames that resolve to any non-public IP, and
-    hostnames that fail to resolve. Bypassed by ``OPENEXTRACT_ALLOW_PRIVATE_URLS``.
-    """
-    if _allow_private_urls():
-        return True
-    if not host:
-        return False
-    bare = host.strip("[]")
-    try:
-        return _is_public_ip(ipaddress.ip_address(bare))
-    except ValueError:
-        pass
-    try:
-        infos = socket.getaddrinfo(bare, None)
-    except OSError:
-        return False
-    if not infos:
-        return False
-    for info in infos:
-        try:
-            resolved = ipaddress.ip_address(info[4][0])
-        except ValueError:
-            return False
-        if not _is_public_ip(resolved):
-            return False
-    return True
-
-
-def _fetch_url(url: str) -> httpx.Response:
-    """Fetch ``url`` with SSRF defenses; validate the host at every redirect hop."""
-    with httpx.Client(follow_redirects=False, timeout=_url_fetch_timeout()) as client:
-        return _fetch_url_with_client(url, client)
-
-
-def _fetch_url_with_client(url: str, client: httpx.Client) -> httpx.Response:
-    """Fetch ``url`` through ``client`` while validating every redirect target."""
-    current = url
-    limit = _max_redirects()
-    for _ in range(limit):
-        host = urlparse(current).hostname
-        if not _is_safe_host(host):
-            raise UrlFetchError(f"Refusing to fetch URL with non-public host: {host!r}")
-        response = client.get(current)
-        if response.is_redirect:
-            location = response.headers.get("location")
-            if not location:
-                raise UrlFetchError(f"Redirect from {current!r} missing Location header")
-            current = str(httpx.URL(current).join(location))
-            continue
-        response.raise_for_status()
-        return response
-    raise UrlFetchError(f"Too many redirects (>{limit})")
-
-
-async def _fetch_url_async(url: str, client: httpx.AsyncClient) -> httpx.Response:
-    """Async URL fetch with the same redirect-by-redirect SSRF validation."""
-    current = url
-    limit = _max_redirects()
-    for _ in range(limit):
-        host = urlparse(current).hostname
-        if not await asyncio.to_thread(_is_safe_host, host):
-            raise UrlFetchError(f"Refusing to fetch URL with non-public host: {host!r}")
-        response = await client.get(current)
-        if response.is_redirect:
-            location = response.headers.get("location")
-            if not location:
-                raise UrlFetchError(f"Redirect from {current!r} missing Location header")
-            current = str(httpx.URL(current).join(location))
-            continue
-        response.raise_for_status()
-        return response
-    raise UrlFetchError(f"Too many redirects (>{limit})")
-
-
-def _response_content_length(response: httpx.Response) -> int | None:
-    raw = response.headers.get("content-length")
-    if raw is None:
-        return None
-    try:
-        value = int(raw)
-    except ValueError:
-        return None
-    return value if value >= 0 else None
-
-
-def _read_response_limited(
-    response: httpx.Response,
-    *,
-    limit: int,
-    source: str,
-) -> bytes:
-    content_length = _response_content_length(response)
-    if content_length is not None and content_length > limit:
-        raise _input_too_large(limit=limit, observed=content_length, source=source)
-
-    chunks: list[bytes] = []
-    total = 0
-    for chunk in response.iter_bytes(chunk_size=_INPUT_READ_CHUNK_SIZE):
-        total += len(chunk)
-        if total > limit:
-            raise _input_too_large(limit=limit, observed=total, source=source)
-        chunks.append(chunk)
-    return b"".join(chunks)
-
-
-async def _read_response_limited_async(
-    response: httpx.Response,
-    *,
-    limit: int,
-    source: str,
-) -> bytes:
-    content_length = _response_content_length(response)
-    if content_length is not None and content_length > limit:
-        raise _input_too_large(limit=limit, observed=content_length, source=source)
-
-    chunks: list[bytes] = []
-    total = 0
-    async for chunk in response.aiter_bytes(chunk_size=_INPUT_READ_CHUNK_SIZE):
-        total += len(chunk)
-        if total > limit:
-            raise _input_too_large(limit=limit, observed=total, source=source)
-        chunks.append(chunk)
-    return b"".join(chunks)
-
-
-def _read_url_with_client(
-    url: str,
-    client: httpx.Client,
-    *,
-    limit: int,
-) -> tuple[bytes, Mapping[str, str]]:
-    """Fetch a URL and stream its final response through the byte cap."""
-    current = url
-    redirect_limit = _max_redirects()
-    for _ in range(redirect_limit):
-        host = urlparse(current).hostname
-        if not _is_safe_host(host):
-            raise UrlFetchError(f"Refusing to fetch URL with non-public host: {host!r}")
-        response = client.send(client.build_request("GET", current), stream=True)
-        try:
-            if response.is_redirect:
-                location = response.headers.get("location")
-                if not location:
-                    raise UrlFetchError(f"Redirect from {current!r} missing Location header")
-                current = str(httpx.URL(current).join(location))
-                continue
-            response.raise_for_status()
-            content = _read_response_limited(
-                response,
-                limit=limit,
-                source=_safe_source_context(current),
-            )
-            return content, dict(response.headers)
-        finally:
-            response.close()
-    raise UrlFetchError(f"Too many redirects (>{redirect_limit})")
-
-
-async def _read_url_with_client_async(
-    url: str,
-    client: httpx.AsyncClient,
-    *,
-    limit: int,
-) -> tuple[bytes, Mapping[str, str]]:
-    """Async counterpart to :func:`_read_url_with_client`."""
-    current = url
-    redirect_limit = _max_redirects()
-    for _ in range(redirect_limit):
-        host = urlparse(current).hostname
-        if not await asyncio.to_thread(_is_safe_host, host):
-            raise UrlFetchError(f"Refusing to fetch URL with non-public host: {host!r}")
-        response = await client.send(client.build_request("GET", current), stream=True)
-        try:
-            if response.is_redirect:
-                location = response.headers.get("location")
-                if not location:
-                    raise UrlFetchError(f"Redirect from {current!r} missing Location header")
-                current = str(httpx.URL(current).join(location))
-                continue
-            response.raise_for_status()
-            content = await _read_response_limited_async(
-                response,
-                limit=limit,
-                source=_safe_source_context(current),
-            )
-            return content, dict(response.headers)
-        finally:
-            await response.aclose()
-    raise UrlFetchError(f"Too many redirects (>{redirect_limit})")
-
-
-def _read_url(url: str, *, limit: int) -> tuple[bytes, Mapping[str, str]]:
-    with httpx.Client(follow_redirects=False, timeout=_url_fetch_timeout()) as client:
-        return _read_url_with_client(url, client, limit=limit)
-
-
-def _media_from_content(
-    file_path: str,
-    content: bytes,
-    headers: Mapping[str, str],
-) -> tuple[bytes, str]:
-    media_type = _get_media_type(file_path)
-    if media_type == _DEFAULT_MEDIA_TYPE:
-        header = headers.get("content-type", "").split(";", 1)[0].strip()
-        if header:
-            media_type = header
-    return content, media_type
-
-
-def _read_from_path(
-    file_path: str,
-    *,
-    max_input_bytes: int,
-    client: httpx.Client | None = None,
-) -> tuple[bytes, str]:
-    """Read bytes from a local path or http(s) URL; return (bytes, media_type)."""
-    if file_path.startswith(_URL_PREFIXES):
-        if client is None:
-            content, headers = _read_url(file_path, limit=max_input_bytes)
-        else:
-            content, headers = _read_url_with_client(
-                file_path,
-                client,
-                limit=max_input_bytes,
-            )
-        return _media_from_content(file_path, content, headers)
-
-    path = Path(file_path)
-    source = _safe_source_context(file_path)
-    size = path.stat().st_size
-    if size > max_input_bytes:
-        raise _input_too_large(limit=max_input_bytes, observed=size, source=source)
-    with path.open("rb") as stream:
-        content = _read_file_like_limited(stream, limit=max_input_bytes, source=source)
-    return content, _get_media_type(file_path)
-
-
-async def _read_from_path_async(
-    file_path: str,
-    client: httpx.AsyncClient | None,
-    *,
-    max_input_bytes: int,
-) -> tuple[bytes, str]:
-    """Async counterpart to :func:`_read_from_path`."""
-    if file_path.startswith(_URL_PREFIXES):
-        if client is None:
-            async with httpx.AsyncClient(
-                follow_redirects=False,
-                timeout=_url_fetch_timeout(),
-            ) as owned_client:
-                content, headers = await _read_url_with_client_async(
-                    file_path,
-                    owned_client,
-                    limit=max_input_bytes,
-                )
-        else:
-            content, headers = await _read_url_with_client_async(
-                file_path,
-                client,
-                limit=max_input_bytes,
-            )
-        return _media_from_content(file_path, content, headers)
-
-    return await asyncio.to_thread(
-        _read_from_path,
-        file_path,
-        max_input_bytes=max_input_bytes,
-    )
-
-
-def _normalize_input(
-    input_file: ExtractionInputLike,
-    media_type: str | None,
-) -> tuple[ResolvedSource, str | None]:
-    """Unwrap an :class:`ExtractionInput` and coerce ``os.PathLike`` sources.
-
-    The effective media type prefers an explicit per-item ``media_type``
-    argument over the value carried on the ``ExtractionInput``. File-like
-    objects (anything exposing ``.read()``) keep precedence over a coincidental
-    ``os.PathLike`` implementation so streams are read, not opened as paths.
-    """
-    if isinstance(input_file, ExtractionInput):
-        if media_type is None:
-            media_type = input_file.media_type
-        input_file = input_file.source
-    if isinstance(input_file, os.PathLike) and not hasattr(input_file, "read"):
-        input_file = os.fspath(input_file)
-    # A PathLike that also exposes ``.read()`` is consumed as a stream, so it is
-    # a valid member of ``ResolvedSource`` at runtime.
-    return cast("ResolvedSource", input_file), media_type
-
-
-def _get_media(
-    input_file: ExtractionInputLike,
-    media_type: str | None = None,
-    *,
-    max_input_bytes: int | None = None,
-    client: httpx.Client | None = None,
-) -> tuple[bytes, str]:
-    """Resolve ``input_file`` to ``(bytes, media_type)``.
-
-    ``str`` and ``os.PathLike`` are treated as a local path or http(s) URL.
-    ``bytes`` and file-like objects (anything with a ``.read()`` method) are
-    passed through. For the latter two, ``media_type`` is required.
-    """
-    input_file, media_type = _normalize_input(input_file, media_type)
-    limit = _resolve_max_input_bytes(max_input_bytes)
-    if isinstance(input_file, str):
-        file_bytes, resolved_type = _read_from_path(
-            input_file,
-            max_input_bytes=limit,
-            client=client,
-        )
-        return file_bytes, media_type or resolved_type
-
-    if isinstance(input_file, bytes):
-        if media_type is None:
-            raise TypeError(_BYTES_MEDIA_TYPE_REQUIRED)
-        return (
-            _enforce_max_input_bytes(input_file, limit=limit, source="bytes input"),
-            media_type,
-        )
-
-    if hasattr(input_file, "read"):
-        if media_type is None:
-            raise TypeError(_BYTES_MEDIA_TYPE_REQUIRED)
-        return (
-            _read_file_like_limited(input_file, limit=limit, source="file-like input"),
-            media_type,
-        )
-
-    raise TypeError(
-        "input_file must be a str path/URL, os.PathLike, bytes, or a file-like "
-        "object with a .read() method."
-    )
-
-
-async def _get_media_async(
-    input_file: ExtractionInputLike,
-    client: httpx.AsyncClient | None = None,
-    media_type: str | None = None,
-    *,
-    max_input_bytes: int | None = None,
-) -> tuple[bytes, str]:
-    """Resolve media without blocking the event loop on disk, DNS, or stream I/O."""
-    input_file, media_type = _normalize_input(input_file, media_type)
-    limit = _resolve_max_input_bytes(max_input_bytes)
-    if isinstance(input_file, str):
-        file_bytes, resolved_type = await _read_from_path_async(
-            input_file,
-            client,
-            max_input_bytes=limit,
-        )
-        return file_bytes, media_type or resolved_type
-
-    if isinstance(input_file, bytes):
-        return _get_media(input_file, media_type=media_type, max_input_bytes=limit)
-
-    if hasattr(input_file, "read"):
-        return await asyncio.to_thread(
-            _get_media,
-            input_file,
-            media_type,
-            max_input_bytes=limit,
-        )
-
-    return _get_media(input_file, media_type=media_type, max_input_bytes=limit)
-
-
-def _install_hint(model: str) -> str:
-    """Return the ``pip install`` command that provides ``model``'s provider."""
-    prefix = model.split(":", 1)[0]
-    extra = _PROVIDER_EXTRAS.get(prefix)
-    target = f"openextract[{extra}]" if extra else "openextract[all]"
-    return f"pip install '{target}'"
-
-
-def _route_model(model: str) -> str:
-    """Route shorthand OpenAI identifiers through the Responses API."""
-    if model.startswith(_OPENAI_PREFIX):
-        return f"{_OPENAI_RESPONSES_PREFIX}{model.removeprefix(_OPENAI_PREFIX)}"
-    return model
-
-
-def _instrumentation_capabilities(
-    instrument: bool | InstrumentationSettings,
-) -> tuple[list[object], dict[str, object]]:
-    """Return version-compatible Pydantic AI instrumentation arguments."""
-    if instrument is False:
-        return [], {}
-
-    from pydantic_ai.models.instrumented import InstrumentationSettings
-
-    if instrument is True:
-        settings = InstrumentationSettings()
-    elif isinstance(instrument, InstrumentationSettings):
-        settings = instrument
-    else:
-        raise TypeError("instrument must be a bool or InstrumentationSettings instance.")
-
-    try:
-        from pydantic_ai.capabilities import Instrumentation
-    except ImportError:  # pragma: no cover - compatibility with older pydantic-ai
-        return [], {"instrument": settings}
-    return [Instrumentation(settings)], {}
-
-
-def _build_agent(
-    schema: type[T],
-    model: str | Model,
-    instructions: str | None,
-    *,
-    model_settings: ModelSettings | None = None,
-    instrument: bool | InstrumentationSettings = False,
-    extra_capabilities: Sequence[object] = (),
-) -> PydanticAgent:
-    """Construct the pydantic_ai Agent, handling the ollama output-type quirk.
-
-    A missing provider SDK surfaces here as ``ImportError`` (pydantic-ai infers
-    the model eagerly); translate it into an actionable
-    :class:`ProviderNotInstalledError`.
-    """
-    try:
-        output_type = schema
-        if isinstance(model, str) and model.startswith("ollama"):
-            from pydantic_ai.output import NativeOutput
-
-            output_type = NativeOutput(schema)
-        capabilities, compatibility_kwargs = _instrumentation_capabilities(instrument)
-        capabilities = [*capabilities, *extra_capabilities]
-        routed_model = _route_model(model) if isinstance(model, str) else model
-        agent_kwargs: dict[str, object] = {
-            "instructions": instructions,
-            "output_type": output_type,
-            "model_settings": model_settings,
-            **compatibility_kwargs,
-        }
-        if capabilities:
-            agent_kwargs["capabilities"] = capabilities
-        return Agent(
-            routed_model,
-            **agent_kwargs,
-        )
-    except ImportError as exc:
-        if isinstance(model, str):
-            message = (
-                f"Model {model!r} needs a provider SDK that is not installed. "
-                f"Install it with: {_install_hint(model)} "
-                f"(or 'pip install openextract[all]'). Original error: {exc}"
-            )
-        else:
-            message = f"The configured model needs a provider SDK that is not installed: {exc}"
-        raise ProviderNotInstalledError(message) from exc
-
-
-def _build_run_inputs(file_bytes: bytes, file_type: str) -> list:
-    """Build the prompt inputs passed to the agent run."""
-    from pydantic_ai import BinaryContent
-
-    return [
-        "Extract the requested information from this document.",
-        BinaryContent(data=file_bytes, media_type=file_type),
-    ]
-
-
-def _model_provider(exc: BaseException) -> str | None:
-    """Return a stable provider identifier when the exception exposes one."""
-    model_name = getattr(exc, "model_name", None)
-    if isinstance(model_name, str) and ":" in model_name:
-        return model_name.split(":", 1)[0]
-
-    for error_type in type(exc).__mro__:
-        module = error_type.__module__
-        for prefix, provider in _PROVIDER_MODULE_NAMES:
-            if module == prefix or module.startswith(f"{prefix}."):
-                return provider
-    return None
-
-
-def _model_status_code(exc: BaseException) -> int | None:
-    """Extract an HTTP status code from common provider exception shapes."""
-    status_code = getattr(exc, "status_code", None)
-    if isinstance(status_code, int) and not isinstance(status_code, bool):
-        return status_code
-
-    response = getattr(exc, "response", None)
-    if response is None:
-        response = getattr(exc, "raw_response", None)
-    response_status = getattr(response, "status_code", None)
-    if isinstance(response_status, int) and not isinstance(response_status, bool):
-        return response_status
-
-    if isinstance(response, Mapping):
-        metadata = response.get("ResponseMetadata", {})
-        if isinstance(metadata, Mapping):
-            metadata_status = metadata.get("HTTPStatusCode")
-            if isinstance(metadata_status, int) and not isinstance(metadata_status, bool):
-                return metadata_status
-
-    code = getattr(exc, "code", None)
-    if isinstance(code, int) and not isinstance(code, bool):
-        return code
-    return None
-
-
-def _parse_retry_after(value: object) -> float | None:
-    """Parse Retry-After seconds or an HTTP date into a non-negative duration."""
-    if not isinstance(value, str | int | float) or isinstance(value, bool):
-        return None
-    try:
-        seconds = float(value)
-    except ValueError:
-        try:
-            retry_at = parsedate_to_datetime(str(value))
-        except (TypeError, ValueError, OverflowError):
-            return None
-        seconds = retry_at.timestamp() - time.time()
-    if not math.isfinite(seconds) or seconds < 0:
-        return None
-    return seconds
-
-
-def _model_retry_after(exc: BaseException) -> float | None:
-    """Extract and parse Retry-After from common provider header containers."""
-    direct_value = getattr(exc, "retry_after", None)
-    if direct_value is not None:
-        parsed_value = _parse_retry_after(direct_value)
-        if parsed_value is not None:
-            return parsed_value
-
-    header_sets: list[Mapping] = []
-    headers = getattr(exc, "headers", None)
-    if isinstance(headers, Mapping):
-        header_sets.append(headers)
-
-    response = getattr(exc, "response", None)
-    if response is None:
-        response = getattr(exc, "raw_response", None)
-    response_headers = getattr(response, "headers", None)
-    if isinstance(response_headers, Mapping):
-        header_sets.append(response_headers)
-    if isinstance(response, Mapping):
-        metadata = response.get("ResponseMetadata", {})
-        if isinstance(metadata, Mapping):
-            metadata_headers = metadata.get("HTTPHeaders", {})
-            if isinstance(metadata_headers, Mapping):
-                header_sets.append(metadata_headers)
-
-    for header_set in header_sets:
-        for key, value in header_set.items():
-            if str(key).lower() == "retry-after":
-                return _parse_retry_after(value)
-    return None
-
-
-def _model_error_name(exc: BaseException) -> str:
-    """Normalize provider class names and structured error codes for policy checks."""
-    names = "".join(error_type.__name__ for error_type in type(exc).__mro__).lower()
-    response = getattr(exc, "response", None)
-    if isinstance(response, Mapping):
-        error = response.get("Error", {})
-        if isinstance(error, Mapping):
-            names += str(error.get("Code", "")).lower()
-    return names.replace("_", "")
-
-
-def _is_transient_model_exception(exc: BaseException, status_code: int | None) -> bool:
-    """Classify only known transient provider failures as retryable."""
-    if status_code is not None:
-        return status_code in _TRANSIENT_HTTP_STATUSES or 500 <= status_code <= 599
-
-    code = getattr(exc, "code", None)
-    if callable(code):
-        grpc_name = getattr(code(), "name", "").lower().replace("_", "")
-        if grpc_name in {
-            "deadlineexceeded",
-            "resourceexhausted",
-            "aborted",
-            "unavailable",
-            "internal",
-        }:
-            return True
-        if grpc_name:
-            return False
-
-    error_name = _model_error_name(exc)
-    if any(name in error_name for name in _PERMANENT_ERROR_NAMES):
-        return False
-    return any(name in error_name for name in _TRANSIENT_ERROR_NAMES)
-
-
-def _map_exception(exc: BaseException) -> ExtractionError:
-    """Translate a low-level exception into the appropriate ExtractionError subclass."""
-    if isinstance(exc, httpx.HTTPStatusError):
-        return UrlFetchError(f"Failed to fetch URL: {exc.response.status_code}")
-    if isinstance(exc, httpx.RequestError):
-        return UrlFetchError(f"Failed to fetch URL: {exc}")
-    if isinstance(exc, ValidationError):
-        return SchemaValidationError(f"Model output did not match schema: {exc}")
-    if _is_model_exception(exc):
-        status_code = _model_status_code(exc)
-        return ModelError(
-            f"Model API error: {exc}",
-            provider=_model_provider(exc),
-            status_code=status_code,
-            retryable=_is_transient_model_exception(exc, status_code),
-            retry_after=_model_retry_after(exc),
-        )
-    return ExtractionError(f"Extraction failed: {exc}")
-
-
-@contextmanager
-def _extraction_errors() -> Iterator[None]:
-    """Map low-level extraction failures to the ``ExtractionError`` hierarchy.
-
-    ``TypeError`` (bad call) and already-mapped ``ExtractionError`` subclasses
-    pass through unchanged; everything else is routed through
-    :func:`_map_exception`.
-    """
-    try:
-        yield
-    except (TypeError, ExtractionError):
-        raise
-    except Exception as e:
-        raise _map_exception(e) from e
-
-
-def _usage_from_result(result) -> Usage:
-    """Build a ``Usage`` from a pydantic-ai run result."""
-    usage_descriptor = getattr(type(result), "usage", None)
-    raw = (
-        result.usage
-        if usage_descriptor is not None and not callable(usage_descriptor)
-        else result.usage()
-    )
-    return Usage(
-        input_tokens=getattr(raw, "input_tokens", 0) or 0,
-        output_tokens=getattr(raw, "output_tokens", 0) or 0,
-        total_tokens=getattr(raw, "total_tokens", 0) or 0,
-    )
-
-
-def total_usage(results: Iterable[ExtractionResult[T]]) -> Usage:
-    """Sum token usage across batch extraction results.
-
-    ``results`` typically comes from :func:`extract_many_with_results` or
-    :func:`extract_many_with_results_async`. Only successful items carry a
-    :class:`Usage`, so totals reflect the successful calls in the batch.
-    """
-    input_tokens = 0
-    output_tokens = 0
-    total_tokens = 0
-    for result in results:
-        input_tokens += result.usage.input_tokens
-        output_tokens += result.usage.output_tokens
-        total_tokens += result.usage.total_tokens
-    return Usage(input_tokens, output_tokens, total_tokens)
-
-
-def _resolve_item(
-    item: ExtractionInputLike,
-    global_media_type: str | None,
-) -> tuple[MediaSource, str | None, str | None]:
-    """Split a batch item into ``(source, effective media type, safe name)``.
-
-    A per-item :class:`ExtractionInput` media type wins over the batch-wide
-    ``media_type`` fallback so heterogeneous inputs can use different types.
-    """
-    if isinstance(item, ExtractionInput):
-        media_type = item.media_type if item.media_type is not None else global_media_type
-        return item.source, media_type, item.name
-    return item, global_media_type, None
-
-
-def _item_source_label(source: MediaSource, name: str | None) -> str | None:
-    """Return a sanitized source label for diagnostics, preferring ``name``.
-
-    URLs and paths are stripped of credentials, query strings, and fragments;
-    raw bytes and file-like inputs have no safe label unless named.
-    """
-    if name is not None:
-        return name
-    if isinstance(source, os.PathLike):
-        source = os.fspath(source)
-    if isinstance(source, str):
-        return _safe_source_context(source)
-    return None
-
-
-def _model_identifier(model: str | Model, agent: object) -> str | None:
-    """Return a stable model identifier for result diagnostics, when known."""
-    if isinstance(model, str):
-        return _route_model(model)
-    name = getattr(model, "model_name", None)
-    if isinstance(name, str):
-        return name
-    agent_model = getattr(agent, "model", None)
-    name = getattr(agent_model, "model_name", None)
-    return name if isinstance(name, str) else None
-
-
-def _resolve_run_inputs(
-    file_bytes: bytes,
-    file_type: str,
-    style_inputs: list[str] | None,
-) -> list:
-    """Use style-specific prompts when present; otherwise pass media directly."""
-    if style_inputs is None:
-        return _build_run_inputs(file_bytes, file_type)
-    return style_inputs
 
 
 @contextmanager
@@ -1160,15 +135,6 @@ async def _prepare_extraction_async(
         yield agent, _resolve_run_inputs(file_bytes, file_type, style_inputs)
 
 
-def _run_extraction(
-    agent: PydanticAgent,
-    inputs: list,
-):
-    """Run a prepared sync extraction and return the raw pydantic-ai result."""
-    with _extraction_errors():
-        return agent.run_sync(inputs)
-
-
 def _extract_once(
     agent: PydanticAgent,
     inputs: list,
@@ -1178,585 +144,21 @@ def _extract_once(
     return cast(T, result.output)
 
 
-async def _run_extraction_async(agent: PydanticAgent, inputs: list):
-    """Run a prepared async extraction with public exception mapping."""
-    with _extraction_errors():
-        return await agent.run(inputs)
-
-
-def _retry_delay(
-    retry_backoff: float,
-    retry_max_backoff: float,
-    attempt: int,
-    retry_after: float | None,
-) -> float:
-    """Return bounded exponential backoff with up to 25% additive jitter."""
-    if retry_after is not None:
-        return min(retry_after, retry_max_backoff)
-    try:
-        delay = retry_backoff * (2**attempt) * (1 + random.uniform(0, 0.25))
-    except OverflowError:
-        return retry_max_backoff
-    return min(delay, retry_max_backoff)
-
-
-def _validate_retry_options(
-    max_retries: object,
-    retry_backoff: object,
-    retry_max_backoff: object,
-) -> None:
-    if isinstance(max_retries, bool) or not isinstance(max_retries, int) or max_retries < 0:
-        raise ValueError("max_retries must be a non-negative integer.")
-    if (
-        isinstance(retry_backoff, bool)
-        or not isinstance(retry_backoff, int | float)
-        or not math.isfinite(retry_backoff)
-        or retry_backoff < 0
-    ):
-        raise ValueError("retry_backoff must be a finite non-negative number of seconds.")
-    if (
-        isinstance(retry_max_backoff, bool)
-        or not isinstance(retry_max_backoff, int | float)
-        or not math.isfinite(retry_max_backoff)
-        or retry_max_backoff < 0
-    ):
-        raise ValueError("retry_max_backoff must be a finite non-negative number of seconds.")
-
-
-def _validate_max_concurrency(max_concurrency: object) -> None:
-    if (
-        isinstance(max_concurrency, bool)
-        or not isinstance(max_concurrency, int)
-        or max_concurrency < 1
-    ):
-        raise ValueError("max_concurrency must be a positive integer.")
-
-
-def _run_with_retries_sync[R](
-    fn: Callable[[], R],
-    *,
+def _oneshot_options(
+    style: ExtractionStyle | str,
     max_retries: int,
     retry_backoff: float,
     retry_max_backoff: float,
-) -> R:
-    """Run ``fn`` until it succeeds or transient retries are exhausted."""
+    max_input_bytes: int | None,
+) -> tuple[ExtractionStyle, int, int, float, float]:
     _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    attempt = 0
-    while True:
-        try:
-            return fn()
-        except ModelError as exc:
-            if not exc.retryable or attempt >= max_retries:
-                raise
-            time.sleep(_retry_delay(retry_backoff, retry_max_backoff, attempt, exc.retry_after))
-            attempt += 1
-
-
-async def _run_with_retries_async[R](
-    fn: Callable[[], Awaitable[R]],
-    *,
-    max_retries: int,
-    retry_backoff: float,
-    retry_max_backoff: float,
-) -> R:
-    """Async counterpart to :func:`_run_with_retries_sync`."""
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    attempt = 0
-    while True:
-        try:
-            return await fn()
-        except ModelError as exc:
-            if not exc.retryable or attempt >= max_retries:
-                raise
-            await asyncio.sleep(
-                _retry_delay(retry_backoff, retry_max_backoff, attempt, exc.retry_after)
-            )
-            attempt += 1
-
-
-def _validate_timeout(value: object, *, name: str) -> float:
-    if (
-        isinstance(value, bool)
-        or not isinstance(value, int | float)
-        or not math.isfinite(value)
-        or value <= 0
-    ):
-        raise ValueError(f"{name} must be a finite positive number of seconds.")
-    return float(value)
-
-
-def _session_model_settings(
-    model_settings: ModelSettings | None,
-    timeout: float | None,
-) -> ModelSettings | None:
-    settings = dict(model_settings) if model_settings is not None else {}
-    if timeout is not None:
-        settings["timeout"] = _validate_timeout(timeout, name="timeout")
-    return cast("ModelSettings | None", settings or None)
-
-
-class _ExtractorSession[T: BaseModel]:
-    """Configuration and output validation shared by sync and async sessions."""
-
-    def __init__(
-        self,
-        schema: type[T],
-        model: str | Model | None,
-        instructions: str | None,
-        *,
-        style: ExtractionStyle | str,
-        agent: PydanticAgent | None,
-        model_settings: ModelSettings | None,
-        timeout: float | None,
-        instrument: bool | InstrumentationSettings,
-        retry_policy: RetryPolicy | None,
-        max_input_bytes: int | None,
-        url_timeout: float | None,
-    ) -> None:
-        resolved_style = normalize_style(style)
-        if agent is not None:
-            if model is not None:
-                raise ValueError("model and agent are mutually exclusive; provide exactly one.")
-            if instructions is not None or model_settings is not None or timeout is not None:
-                raise ValueError(
-                    "instructions, model_settings, and timeout must be configured "
-                    "on an injected agent."
-                )
-            if instrument is not False:
-                raise ValueError("instrument must be configured on an injected agent.")
-            if resolved_style is not ExtractionStyle.DIRECT:
-                raise ValueError("style other than 'direct' cannot be used with an injected agent.")
-            configured_agent = agent
-            session_settings = None
-        else:
-            if model is None:
-                raise TypeError("model is required unless agent is provided.")
-            session_settings = _session_model_settings(model_settings, timeout)
-            if resolved_style is ExtractionStyle.DIRECT:
-                configured_agent = _build_agent(
-                    schema,
-                    model,
-                    instructions,
-                    model_settings=session_settings,
-                    instrument=instrument,
-                )
-            else:
-                configured_agent = None
-
-        if retry_policy is None:
-            retry_policy = RetryPolicy()
-        elif not isinstance(retry_policy, RetryPolicy):
-            raise TypeError("retry_policy must be a RetryPolicy instance.")
-
-        self._schema = schema
-        self._model = model
-        self._instructions = instructions
-        self._style = resolved_style
-        self._model_settings = session_settings
-        self._instrument = instrument
-        self._agent = configured_agent
-        self._retry_policy = retry_policy
-        self._max_input_bytes = _resolve_max_input_bytes(max_input_bytes)
-        self._url_timeout = (
-            _url_fetch_timeout()
-            if url_timeout is None
-            else _validate_timeout(url_timeout, name="url_timeout")
-        )
-        self._entered = False
-        self._closed = False
-        self._style_workspace: tempfile.TemporaryDirectory[str] | None = None
-        self._style_run_index = 0
-
-    def _validate_output(self, output: object) -> T:
-        with _extraction_errors():
-            return self._schema.model_validate(output)
-
-    def _ensure_enterable(self, class_name: str) -> None:
-        if self._closed:
-            raise RuntimeError(f"{class_name} is closed and cannot be reused.")
-        if self._entered:
-            raise RuntimeError(f"{class_name} is already entered.")
-
-    def _ensure_open(self, class_name: str) -> None:
-        if not self._entered:
-            raise RuntimeError(f"{class_name} must be used as a context manager before extraction.")
-
-    def _enter_style_workspace(self) -> None:
-        """Create the session workspace and its agent for search/code styles.
-
-        The agent (and its provider HTTP client) is built once per session and
-        lives until the session closes, matching the direct-style lifecycle.
-        """
-        if self._style is ExtractionStyle.DIRECT:
-            return
-        assert self._model is not None
-        self._style_workspace = tempfile.TemporaryDirectory(prefix="openextract-")
-        with _extraction_errors():
-            self._agent = _build_agent(
-                self._schema,
-                self._model,
-                self._instructions,
-                model_settings=self._model_settings,
-                instrument=self._instrument,
-                extra_capabilities=style_capabilities(
-                    self._style, Path(self._style_workspace.name)
-                ),
-            )
-
-    def _discard_style_state(self) -> None:
-        """Drop the style agent and remove the workspace owned by the session."""
-        if self._style_workspace is not None:
-            self._style_workspace.cleanup()
-            self._style_workspace = None
-            self._agent = None
-
-    @contextmanager
-    def _style_document(self, file_bytes: bytes, file_type: str) -> Iterator[list]:
-        """Materialize one document in the session workspace for a single call.
-
-        Each call gets its own subdirectory so concurrent async extractions
-        never collide. The subdirectory is removed when the call finishes; the
-        workspace itself lives until the session closes, so documents persist
-        across retries within a call.
-        """
-        assert self._style_workspace is not None
-        self._style_run_index += 1
-        run_dir = Path(self._style_workspace.name) / f"run{self._style_run_index}"
-        run_dir.mkdir()
-        try:
-            filename = materialize_text_document(run_dir, file_bytes, file_type, style=self._style)
-            yield style_run_inputs(self._style, f"{run_dir.name}/{filename}")
-        finally:
-            shutil.rmtree(run_dir, ignore_errors=True)
-
-    @contextmanager
-    def _session_agent_inputs(
-        self, file_bytes: bytes, file_type: str
-    ) -> Iterator[tuple[PydanticAgent, list]]:
-        """Pair the session agent with per-call run inputs for one extraction."""
-        assert self._agent is not None
-        if self._style is ExtractionStyle.DIRECT:
-            yield self._agent, _build_run_inputs(file_bytes, file_type)
-            return
-        with self._style_document(file_bytes, file_type) as inputs:
-            yield self._agent, inputs
-
-
-class Extractor(_ExtractorSession[T]):
-    """Reusable synchronous extraction session.
-
-    An ``Extractor`` is bound to the thread that enters it and is not
-    thread-safe. Use one session per thread and close it deterministically with
-    a ``with`` block. ``search``/``code`` sessions build one harness agent and
-    temporary workspace on enter and remove both on close.
-    """
-
-    def __init__(
-        self,
-        schema: type[T],
-        model: str | Model | None = None,
-        instructions: str | None = None,
-        *,
-        style: ExtractionStyle | str = "direct",
-        agent: PydanticAgent | None = None,
-        model_settings: ModelSettings | None = None,
-        timeout: float | None = None,
-        instrument: bool | InstrumentationSettings = False,
-        retry_policy: RetryPolicy | None = None,
-        max_input_bytes: int | None = None,
-        url_timeout: float | None = None,
-    ) -> None:
-        super().__init__(
-            schema,
-            model,
-            instructions,
-            style=style,
-            agent=agent,
-            model_settings=model_settings,
-            timeout=timeout,
-            instrument=instrument,
-            retry_policy=retry_policy,
-            max_input_bytes=max_input_bytes,
-            url_timeout=url_timeout,
-        )
-        self._client: httpx.Client | None = None
-        self._runner: asyncio.Runner | None = None
-        self._thread_id: int | None = None
-
-    def __enter__(self) -> Extractor[T]:
-        self._ensure_enterable(type(self).__name__)
-        try:
-            asyncio.get_running_loop()
-        except RuntimeError:
-            pass
-        else:
-            raise RuntimeError(
-                "Extractor cannot be entered from a running event loop; use AsyncExtractor instead."
-            )
-
-        client = httpx.Client(follow_redirects=False, timeout=self._url_timeout)
-        # Keep the session loop private so entering an Extractor does not replace
-        # or clear the caller's process-wide current event loop.
-        runner = asyncio.Runner(loop_factory=asyncio.new_event_loop)
-        runner.__enter__()
-        try:
-            self._enter_style_workspace()
-            assert self._agent is not None
-            runner.run(self._agent.__aenter__())
-        except BaseException:
-            self._discard_style_state()
-            client.close()
-            runner.close()
-            raise
-
-        self._client = client
-        self._runner = runner
-        self._thread_id = threading.get_ident()
-        self._entered = True
-        return self
-
-    def __exit__(self, *exc_info: object) -> bool:
-        return self._close(exc_info)
-
-    def _close(self, exc_info: tuple[object, ...]) -> bool:
-        if not self._entered:
-            self._closed = True
-            return False
-        if self._thread_id != threading.get_ident():
-            raise RuntimeError("Extractor can only be closed from the thread that entered it.")
-        assert self._runner is not None
-        assert self._client is not None
-        assert self._agent is not None
-        suppressed = False
-        try:
-            suppressed = bool(self._runner.run(self._agent.__aexit__(*exc_info)))
-        finally:
-            self._client.close()
-            self._runner.close()
-            self._discard_style_state()
-            self._entered = False
-            self._closed = True
-            self._client = None
-            self._runner = None
-        return suppressed
-
-    def close(self) -> None:
-        """Close the owned agent/provider and input HTTP client."""
-        self._close((None, None, None))
-
-    def _ensure_sync_open(self) -> httpx.Client:
-        self._ensure_open(type(self).__name__)
-        if self._thread_id != threading.get_ident():
-            raise RuntimeError("Extractor can only be used from the thread that entered it.")
-        assert self._client is not None
-        return self._client
-
-    def _run_agent(self, agent: PydanticAgent, inputs: list):
-        assert self._runner is not None
-        return self._runner.run(_run_extraction_async(agent, inputs))
-
-    @contextmanager
-    def _prepare_session_extraction(
-        self,
-        input_file: ExtractionInputLike,
-        media_type: str | None,
-    ) -> Iterator[tuple[PydanticAgent, list]]:
-        """Resolve media and yield ``(agent, inputs)`` for one session call."""
-        client = self._ensure_sync_open()
-        with _extraction_errors():
-            file_bytes, file_type = _get_media(
-                input_file,
-                media_type=media_type,
-                max_input_bytes=self._max_input_bytes,
-                client=client,
-            )
-        with self._session_agent_inputs(file_bytes, file_type) as prepared:
-            yield prepared
-
-    def _run_session_retries[R](self, fn: Callable[[], R]) -> R:
-        return _run_with_retries_sync(
-            fn,
-            max_retries=self._retry_policy.max_retries,
-            retry_backoff=self._retry_policy.backoff,
-            retry_max_backoff=self._retry_policy.max_backoff,
-        )
-
-    def extract(
-        self,
-        input_file: ExtractionInputLike,
-        *,
-        media_type: str | None = None,
-    ) -> T:
-        """Extract one input using the session's reusable agent and clients."""
-        with self._prepare_session_extraction(input_file, media_type) as (agent, inputs):
-            return self._run_session_retries(
-                lambda: self._validate_output(self._run_agent(agent, inputs).output)
-            )
-
-    def extract_with_usage(
-        self,
-        input_file: ExtractionInputLike,
-        *,
-        media_type: str | None = None,
-    ) -> tuple[T, Usage]:
-        """Extract one input and return its successful-call token usage."""
-        with self._prepare_session_extraction(input_file, media_type) as (agent, inputs):
-
-            def _once() -> tuple[T, Usage]:
-                result = self._run_agent(agent, inputs)
-                return self._validate_output(result.output), _usage_from_result(result)
-
-            return self._run_session_retries(_once)
-
-
-class AsyncExtractor(_ExtractorSession[T]):
-    """Reusable async extraction session bound to one event loop.
-
-    ``search``/``code`` sessions build one harness agent and temporary
-    workspace on enter and remove both on close; concurrent calls each write
-    their document into a private workspace subdirectory.
-    """
-
-    def __init__(
-        self,
-        schema: type[T],
-        model: str | Model | None = None,
-        instructions: str | None = None,
-        *,
-        style: ExtractionStyle | str = "direct",
-        agent: PydanticAgent | None = None,
-        model_settings: ModelSettings | None = None,
-        timeout: float | None = None,
-        instrument: bool | InstrumentationSettings = False,
-        retry_policy: RetryPolicy | None = None,
-        max_input_bytes: int | None = None,
-        url_timeout: float | None = None,
-    ) -> None:
-        super().__init__(
-            schema,
-            model,
-            instructions,
-            style=style,
-            agent=agent,
-            model_settings=model_settings,
-            timeout=timeout,
-            instrument=instrument,
-            retry_policy=retry_policy,
-            max_input_bytes=max_input_bytes,
-            url_timeout=url_timeout,
-        )
-        self._client: httpx.AsyncClient | None = None
-        self._loop: asyncio.AbstractEventLoop | None = None
-
-    async def __aenter__(self) -> AsyncExtractor[T]:
-        self._ensure_enterable(type(self).__name__)
-        client = httpx.AsyncClient(follow_redirects=False, timeout=self._url_timeout)
-        try:
-            self._enter_style_workspace()
-            assert self._agent is not None
-            await self._agent.__aenter__()
-        except BaseException:
-            self._discard_style_state()
-            await client.aclose()
-            raise
-        self._client = client
-        self._loop = asyncio.get_running_loop()
-        self._entered = True
-        return self
-
-    async def __aexit__(self, *exc_info: object) -> bool:
-        return await self._close(exc_info)
-
-    async def _close(self, exc_info: tuple[object, ...]) -> bool:
-        if not self._entered:
-            self._closed = True
-            return False
-        if self._loop is not asyncio.get_running_loop():
-            raise RuntimeError(
-                "AsyncExtractor can only be closed from the event loop that entered it."
-            )
-        assert self._client is not None
-        assert self._agent is not None
-        suppressed = False
-        try:
-            suppressed = bool(await self._agent.__aexit__(*exc_info))
-        finally:
-            await self._client.aclose()
-            self._discard_style_state()
-            self._entered = False
-            self._closed = True
-            self._client = None
-            self._loop = None
-        return suppressed
-
-    async def aclose(self) -> None:
-        """Close the owned agent/provider and input HTTP client."""
-        await self._close((None, None, None))
-
-    def _ensure_async_open(self) -> httpx.AsyncClient:
-        self._ensure_open(type(self).__name__)
-        if self._loop is not asyncio.get_running_loop():
-            raise RuntimeError(
-                "AsyncExtractor can only be used from the event loop that entered it."
-            )
-        assert self._client is not None
-        return self._client
-
-    @asynccontextmanager
-    async def _prepare_session_extraction(
-        self,
-        input_file: ExtractionInputLike,
-        media_type: str | None,
-    ) -> AsyncIterator[tuple[PydanticAgent, list]]:
-        """Resolve media and yield ``(agent, inputs)`` for one session call."""
-        client = self._ensure_async_open()
-        with _extraction_errors():
-            file_bytes, file_type = await _get_media_async(
-                input_file,
-                client,
-                media_type=media_type,
-                max_input_bytes=self._max_input_bytes,
-            )
-        with self._session_agent_inputs(file_bytes, file_type) as prepared:
-            yield prepared
-
-    async def _run_session_retries[R](self, fn: Callable[[], Awaitable[R]]) -> R:
-        return await _run_with_retries_async(
-            fn,
-            max_retries=self._retry_policy.max_retries,
-            retry_backoff=self._retry_policy.backoff,
-            retry_max_backoff=self._retry_policy.max_backoff,
-        )
-
-    async def extract(
-        self,
-        input_file: ExtractionInputLike,
-        *,
-        media_type: str | None = None,
-    ) -> T:
-        """Extract one input using the session's reusable agent and clients."""
-        async with self._prepare_session_extraction(input_file, media_type) as (agent, inputs):
-
-            async def _once() -> T:
-                result = await _run_extraction_async(agent, inputs)
-                return self._validate_output(result.output)
-
-            return await self._run_session_retries(_once)
-
-    async def extract_with_usage(
-        self,
-        input_file: ExtractionInputLike,
-        *,
-        media_type: str | None = None,
-    ) -> tuple[T, Usage]:
-        """Extract one input and return its successful-call token usage."""
-        async with self._prepare_session_extraction(input_file, media_type) as (agent, inputs):
-
-            async def _once() -> tuple[T, Usage]:
-                result = await _run_extraction_async(agent, inputs)
-                return self._validate_output(result.output), _usage_from_result(result)
-
-            return await self._run_session_retries(_once)
+    return (
+        normalize_style(style),
+        _resolve_max_input_bytes(max_input_bytes),
+        max_retries,
+        retry_backoff,
+        retry_max_backoff,
+    )
 
 
 def extract(
@@ -1813,9 +215,9 @@ def extract(
         ValueError: If ``style`` is invalid or ``search``/``code`` is used with
             a non-text document.
     """
-    style = normalize_style(style)
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    limit = _resolve_max_input_bytes(max_input_bytes)
+    style, limit, max_retries, retry_backoff, retry_max_backoff = _oneshot_options(
+        style, max_retries, retry_backoff, retry_max_backoff, max_input_bytes
+    )
     with _prepare_extraction(
         schema,
         model,
@@ -1851,9 +253,9 @@ def extract_with_usage(
     Same retry semantics as :func:`extract`. Returns a :class:`Usage` describing
     the tokens consumed by the successful model call.
     """
-    style = normalize_style(style)
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    limit = _resolve_max_input_bytes(max_input_bytes)
+    style, limit, max_retries, retry_backoff, retry_max_backoff = _oneshot_options(
+        style, max_retries, retry_backoff, retry_max_backoff, max_input_bytes
+    )
     with _prepare_extraction(
         schema,
         model,
@@ -1890,9 +292,9 @@ async def extract_with_usage_async(
     retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
 ) -> tuple[T, Usage]:
     """Async sibling of :func:`extract_with_usage`; returns ``(output, Usage)``."""
-    style = normalize_style(style)
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    limit = _resolve_max_input_bytes(max_input_bytes)
+    style, limit, max_retries, retry_backoff, retry_max_backoff = _oneshot_options(
+        style, max_retries, retry_backoff, retry_max_backoff, max_input_bytes
+    )
     async with _prepare_extraction_async(
         schema,
         model,
@@ -1929,9 +331,9 @@ async def extract_async(
     retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
 ) -> T:
     """Async sibling of :func:`extract`; uses ``Agent.run`` instead of ``run_sync``."""
-    style = normalize_style(style)
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    limit = _resolve_max_input_bytes(max_input_bytes)
+    style, limit, max_retries, retry_backoff, retry_max_backoff = _oneshot_options(
+        style, max_retries, retry_backoff, retry_max_backoff, max_input_bytes
+    )
     async with _prepare_extraction_async(
         schema,
         model,
@@ -1954,789 +356,53 @@ async def extract_async(
         )
 
 
-async def _run_with_shared_agent(
-    agent: PydanticAgent,
-    inputs: list,
-) -> object:
-    """Run prepared inputs through a pre-built shared ``Agent``.
-
-    Mirrors ``extract_async``'s error mapping so callers get the same
-    ``ExtractionError`` subclasses as the per-item path. Returns the validated
-    schema instance.
-    """
-    result = await _run_extraction_async(agent, inputs)
-    return result.output
-
-
-async def _run_with_shared_agent_result(
-    agent: PydanticAgent,
-    inputs: list,
-) -> object:
-    """Run prepared inputs through a shared ``Agent`` and return the raw result.
-
-    The raw pydantic-ai result exposes ``.output`` and ``.usage()`` so the rich
-    batch path can build :class:`ExtractionResult` diagnostics.
-    """
-    return await _run_extraction_async(agent, inputs)
-
-
-async def _cancel_tasks(tasks: Iterable[asyncio.Task[object]]) -> None:
-    """Cancel and await every task so no batch work outlives its caller."""
-    task_list = list(tasks)
-    for task in task_list:
-        task.cancel()
-    if task_list:
-        await asyncio.gather(*task_list, return_exceptions=True)
-
-
-async def _iter_extractions(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None,
-    max_concurrency: int,
-    return_exceptions: bool,
-    media_type: str | None,
-    max_input_bytes: int,
-    max_retries: int,
-    retry_backoff: float,
-    retry_max_backoff: float,
-    rich: bool = False,
-    style: ExtractionStyle = ExtractionStyle.DIRECT,
-) -> AsyncIterator[tuple[int, T | ExtractionResult[T] | Exception]]:
-    """Yield indexed batch results in completion order with bounded work.
-
-    When ``rich`` is true each successful item is yielded as an
-    :class:`ExtractionResult[T]`; otherwise the bare schema instance is yielded.
-    """
-    file_iterator = iter(input_files)
-    try:
-        first_item = next(file_iterator)
-    except StopIteration:
-        return
-
-    # Building the Agent (and its provider HTTP client) is ~32 ms; sharing one
-    # across the batch saves ~32 ms × (N-1) per call. The Agent is stateless
-    # between runs and stays inside this event loop, so this is safe. Search and
-    # code styles bind capabilities to a per-item workspace, so those items
-    # each get their own agent.
-    shared_agent = (
-        _build_agent(schema, model, instructions) if style is ExtractionStyle.DIRECT else None
-    )
-    stop = asyncio.Event()
-    pending: dict[asyncio.Task[object], int] = {}
-    next_index = 0
-    exhausted = False
-
-    async with httpx.AsyncClient(
-        follow_redirects=False,
-        timeout=_url_fetch_timeout(),
-    ) as client:
-
-        async def _run_item(item: ExtractionInputLike) -> object:
-            source, item_media_type, name = _resolve_item(item, media_type)
-            started = time.perf_counter()
-            attempts = 0
-            try:
-                with _extraction_errors():
-                    file_bytes, file_type = await _get_media_async(
-                        source,
-                        client,
-                        media_type=item_media_type,
-                        max_input_bytes=max_input_bytes,
-                    )
-                with prepared_style_run(style, file_bytes, file_type) as (
-                    capabilities,
-                    style_inputs,
-                ):
-                    inputs = _resolve_run_inputs(file_bytes, file_type, style_inputs)
-                    if shared_agent is None:
-                        with _extraction_errors():
-                            run_agent = _build_agent(
-                                schema,
-                                model,
-                                instructions,
-                                extra_capabilities=capabilities,
-                            )
-                    else:
-                        run_agent = shared_agent
-
-                    async def _once() -> object:
-                        nonlocal attempts
-                        attempts += 1
-                        # A sibling may have failed while this item was being prepared
-                        # or waiting to retry. Do not begin another model call afterward.
-                        if stop.is_set():
-                            raise asyncio.CancelledError
-                        if rich:
-                            return await _run_with_shared_agent_result(run_agent, inputs)
-                        return await _run_with_shared_agent(run_agent, inputs)
-
-                    value = await _run_with_retries_async(
-                        _once,
-                        max_retries=max_retries,
-                        retry_backoff=retry_backoff,
-                        retry_max_backoff=retry_max_backoff,
-                    )
-                if rich:
-                    raw_result = cast(Any, value)
-                    return ExtractionResult(
-                        output=cast(T, raw_result.output),
-                        usage=_usage_from_result(raw_result),
-                        attempts=attempts,
-                        duration=time.perf_counter() - started,
-                        model=_model_identifier(model, run_agent),
-                        media_type=item_media_type,
-                        source=_item_source_label(source, name),
-                        warnings=(),
-                    )
-                return value
-            except Exception:
-                if not return_exceptions:
-                    stop.set()
-                raise
-
-        def _schedule(item: ExtractionInputLike) -> None:
-            nonlocal next_index
-            task = asyncio.create_task(_run_item(item))
-            pending[task] = next_index
-            next_index += 1
-
-        def _fill_slots() -> None:
-            nonlocal exhausted
-            while len(pending) < max_concurrency and not exhausted:
-                if next_index == 0:
-                    item = first_item
-                else:
-                    try:
-                        item = next(file_iterator)
-                    except StopIteration:
-                        exhausted = True
-                        break
-                _schedule(item)
-
-        try:
-            _fill_slots()
-            while pending:
-                done, _ = await asyncio.wait(pending, return_when=asyncio.FIRST_COMPLETED)
-                completed: list[tuple[int, T | ExtractionResult[T] | Exception]] = []
-                failures: list[Exception] = []
-                child_cancelled = False
-
-                # Stable index ordering makes simultaneous completions deterministic.
-                for task in sorted(done, key=pending.__getitem__):
-                    index = pending.pop(task)
-                    if task.cancelled():
-                        child_cancelled = True
-                        continue
-                    try:
-                        result = task.result()
-                    except Exception as exc:
-                        if return_exceptions:
-                            completed.append((index, exc))
-                        else:
-                            failures.append(exc)
-                    else:
-                        completed.append((index, cast("T | ExtractionResult[T]", result)))
-
-                if failures:
-                    await _cancel_tasks(pending)
-                    pending.clear()
-                    raise failures[0]
-                if child_cancelled:
-                    raise asyncio.CancelledError
-
-                # Refill only after every completion has been checked for a
-                # fail-fast error. Pending tasks therefore stay O(concurrency).
-                _fill_slots()
-                for indexed_result in completed:
-                    yield indexed_result
-        finally:
-            await _cancel_tasks(pending)
-            pending.clear()
-
-
-async def _gather_extractions(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None,
-    max_concurrency: int,
-    return_exceptions: bool,
-    media_type: str | None,
-    max_input_bytes: int | None,
-    max_retries: int,
-    retry_backoff: float,
-    retry_max_backoff: float,
-    rich: bool = False,
-    style: ExtractionStyle = ExtractionStyle.DIRECT,
-) -> list:
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    _validate_max_concurrency(max_concurrency)
-    limit = _resolve_max_input_bytes(max_input_bytes)
-    indexed_results = [
-        item
-        async for item in _iter_extractions(
-            schema,
-            model,
-            input_files,
-            instructions,
-            max_concurrency,
-            return_exceptions,
-            media_type,
-            limit,
-            max_retries,
-            retry_backoff,
-            retry_max_backoff,
-            rich,
-            style,
-        )
-    ]
-    indexed_results.sort(key=lambda item: item[0])
-    return [result for _, result in indexed_results]
-
-
-@overload
-def extract_many(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[False] = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[T]: ...
-
-
-@overload
-def extract_many(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[True],
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[T | Exception]: ...
-
-
-@overload
-def extract_many(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[T | Exception]: ...
-
-
-def extract_many(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list:
-    """Run :func:`extract` over many inputs concurrently from sync code.
-
-    Each item may be a raw path/URL/``bytes``/file-like ``os.PathLike`` or an
-    :class:`ExtractionInput` carrying a per-item ``media_type``.
-
-    Args:
-        schema: A Pydantic model class defining the expected output structure.
-        model: The model identifier.
-        input_files: Iterable of paths, URLs, ``os.PathLike``, bytes,
-            file-like objects, or :class:`ExtractionInput` items.
-        instructions: Optional natural-language guidance.
-        media_type: Optional MIME type applied to every item that does not carry
-            its own. Required when ``input_files`` contains ``bytes`` or
-            file-like objects without a per-item ``media_type``; optional
-            override for path/URL items.
-        max_input_bytes: Per-item byte limit. ``None`` uses
-            ``OPENEXTRACT_MAX_INPUT_BYTES`` or the 50 MiB default.
-        max_concurrency: Maximum number of in-flight extractions.
-        return_exceptions: If True, exceptions are returned in-place instead of
-            raised (mirrors :func:`asyncio.gather`).
-        max_retries: Per-item retries after ``ModelError`` (same semantics as
-            :func:`extract`).
-        retry_backoff: Base backoff seconds between per-item retries.
-        retry_max_backoff: Maximum per-item retry delay in seconds.
-
-    Returns:
-        A list of results (or exceptions, when ``return_exceptions=True``) in
-        input order.
-
-    Raises:
-        ValueError: If ``max_concurrency`` is less than 1, ``max_retries`` is
-            negative, or a backoff value is negative or non-finite.
-        RuntimeError: If called from a running event loop. Use
-            :func:`extract_many_async` in async code instead.
-    """
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    _validate_max_concurrency(max_concurrency)
-    try:
-        asyncio.get_running_loop()
-    except RuntimeError:
-        pass
-    else:
-        raise RuntimeError(
-            "extract_many() cannot be called from a running event loop; "
-            "use await extract_many_async(...) instead."
-        )
-    return asyncio.run(
-        _gather_extractions(
-            schema,
-            model,
-            input_files,
-            instructions,
-            max_concurrency,
-            return_exceptions,
-            media_type,
-            max_input_bytes,
-            max_retries,
-            retry_backoff,
-            retry_max_backoff,
-            style=normalize_style(style),
-        )
-    )
-
-
-@overload
-async def extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[False] = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[T]: ...
-
-
-@overload
-async def extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[True],
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[T | Exception]: ...
-
-
-@overload
-async def extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[T | Exception]: ...
-
-
-async def extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list:
-    """Async sibling of :func:`extract_many`."""
-    return await _gather_extractions(
-        schema,
-        model,
-        input_files,
-        instructions,
-        max_concurrency,
-        return_exceptions,
-        media_type,
-        max_input_bytes,
-        max_retries,
-        retry_backoff,
-        retry_max_backoff,
-        style=normalize_style(style),
-    )
-
-
-@overload
-def iter_extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[False] = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> AsyncIterator[tuple[int, T]]: ...
-
-
-@overload
-def iter_extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[True],
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> AsyncIterator[tuple[int, T | Exception]]: ...
-
-
-@overload
-def iter_extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> AsyncIterator[tuple[int, T | Exception]]: ...
-
-
-def iter_extract_many_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> AsyncIterator[tuple[int, T | Exception]]:
-    """Stream ``(input_index, result)`` pairs in completion order.
-
-    Unlike :func:`extract_many_async`, this API does not eagerly consume
-    ``input_files`` and does not wait for the complete batch before yielding.
-    At most ``max_concurrency`` items are scheduled at once. If
-    ``return_exceptions`` is true, item failures are yielded as the result;
-    otherwise the first failure cancels and awaits all outstanding work.
-    ``max_input_bytes`` applies the same per-item cap as the list APIs.
-
-    The function itself is synchronous because it returns an async iterator::
-
-        async for index, result in iter_extract_many_async(...):
-            ...
-    """
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    _validate_max_concurrency(max_concurrency)
-    limit = _resolve_max_input_bytes(max_input_bytes)
-    # The shared generator is typed for the richest yield (ExtractionResult);
-    # without ``rich`` it only ever yields ``T`` or an exception, so narrow it.
-    return cast(
-        AsyncIterator[tuple[int, T | Exception]],
-        _iter_extractions(
-            schema,
-            model,
-            input_files,
-            instructions,
-            max_concurrency,
-            return_exceptions,
-            media_type,
-            limit,
-            max_retries,
-            retry_backoff,
-            retry_max_backoff,
-            style=normalize_style(style),
-        ),
-    )
-
-
-@overload
-def extract_many_with_results(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[False] = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[ExtractionResult[T]]: ...
-
-
-@overload
-def extract_many_with_results(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[True],
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[ExtractionResult[T] | Exception]: ...
-
-
-@overload
-def extract_many_with_results(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[ExtractionResult[T] | Exception]: ...
-
-
-def extract_many_with_results(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list:
-    """Run a batch and return per-item :class:`ExtractionResult` diagnostics.
-
-    Mirrors :func:`extract_many` (same arguments, ordering, concurrency, and
-    retry semantics) but returns richer results carrying token usage, attempt
-    counts, timing, model/media metadata, and a sanitized source label. Use
-    :func:`total_usage` to aggregate token usage across the returned results.
-
-    Args:
-        schema: A Pydantic model class defining the expected output structure.
-        model: The model identifier.
-        input_files: Iterable of paths, URLs, ``os.PathLike``, bytes,
-            file-like objects, or :class:`ExtractionInput` items.
-        instructions: Optional natural-language guidance.
-        media_type: Optional MIME type applied to every item that does not carry
-            its own.
-        max_input_bytes: Per-item byte limit. ``None`` uses
-            ``OPENEXTRACT_MAX_INPUT_BYTES`` or the 50 MiB default.
-        max_concurrency: Maximum number of in-flight extractions.
-        return_exceptions: If True, per-item exceptions are returned in-place
-            instead of raised.
-        max_retries: Per-item retries after ``ModelError``.
-        retry_backoff: Base backoff seconds between per-item retries.
-        retry_max_backoff: Maximum per-item retry delay in seconds.
-
-    Returns:
-        A list of ``ExtractionResult`` (or ``Exception`` when
-        ``return_exceptions=True``) in input order.
-
-    Raises:
-        ValueError: If ``max_concurrency`` is less than 1, ``max_retries`` is
-            negative, or a backoff value is negative or non-finite.
-        RuntimeError: If called from a running event loop. Use
-            :func:`extract_many_with_results_async` in async code instead.
-    """
-    _validate_retry_options(max_retries, retry_backoff, retry_max_backoff)
-    _validate_max_concurrency(max_concurrency)
-    try:
-        asyncio.get_running_loop()
-    except RuntimeError:
-        pass
-    else:
-        raise RuntimeError(
-            "extract_many_with_results() cannot be called from a running event loop; "
-            "use await extract_many_with_results_async(...) instead."
-        )
-    return asyncio.run(
-        _gather_extractions(
-            schema,
-            model,
-            input_files,
-            instructions,
-            max_concurrency,
-            return_exceptions,
-            media_type,
-            max_input_bytes,
-            max_retries,
-            retry_backoff,
-            retry_max_backoff,
-            rich=True,
-            style=normalize_style(style),
-        )
-    )
-
-
-@overload
-async def extract_many_with_results_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[False] = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[ExtractionResult[T]]: ...
-
-
-@overload
-async def extract_many_with_results_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: Literal[True],
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[ExtractionResult[T] | Exception]: ...
-
-
-@overload
-async def extract_many_with_results_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list[ExtractionResult[T] | Exception]: ...
-
-
-async def extract_many_with_results_async(
-    schema: type[T],
-    model: str | Model,
-    input_files: Iterable[ExtractionInputLike],
-    instructions: str | None = None,
-    *,
-    style: ExtractionStyle | str = "direct",
-    media_type: str | None = None,
-    max_input_bytes: int | None = None,
-    max_concurrency: int = 5,
-    return_exceptions: bool = False,
-    max_retries: int = 0,
-    retry_backoff: float = 1.0,
-    retry_max_backoff: float = _DEFAULT_RETRY_MAX_BACKOFF,
-) -> list:
-    """Async sibling of :func:`extract_many_with_results`."""
-    return await _gather_extractions(
-        schema,
-        model,
-        input_files,
-        instructions,
-        max_concurrency,
-        return_exceptions,
-        media_type,
-        max_input_bytes,
-        max_retries,
-        retry_backoff,
-        retry_max_backoff,
-        rich=True,
-        style=normalize_style(style),
-    )
+__all__ = [
+    "Agent",
+    "Extractor",
+    "AsyncExtractor",
+    "RetryPolicy",
+    "ExtractionInput",
+    "ExtractionResult",
+    "Usage",
+    "extract",
+    "extract_async",
+    "extract_with_usage",
+    "extract_with_usage_async",
+    "extract_many",
+    "extract_many_async",
+    "iter_extract_many_async",
+    "extract_many_with_results",
+    "extract_many_with_results_async",
+    "total_usage",
+    "_build_agent",
+    "_extract_once",
+    "_fetch_url",
+    "_fetch_url_async",
+    "_get_media",
+    "_get_media_async",
+    "_get_media_type",
+    "_install_hint",
+    "_is_public_ip",
+    "_is_safe_host",
+    "_is_transient_model_exception",
+    "_item_source_label",
+    "_map_exception",
+    "_max_redirects",
+    "_model_identifier",
+    "_model_retry_after",
+    "_model_status_code",
+    "_parse_retry_after",
+    "_prepare_extraction",
+    "_read_from_path",
+    "_read_url_with_client",
+    "_resolve_item",
+    "_resolve_max_input_bytes",
+    "_resolve_run_inputs",
+    "_retry_delay",
+    "_route_model",
+    "_run_extraction",
+    "_run_with_shared_agent",
+    "_run_with_shared_agent_result",
+    "_safe_source_context",
+    "_url_fetch_timeout",
+]
